backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.31.0
PyMySQL==1.0.3
channels==4.1.0
channels-redis==4.0.0
python-dotenv==1.0.0